requirements.txt
beautifulsoup4==4.12.3
python-dotenv==1.0.1
requests==2.32.3